Re-arrange and update the file list stories.
Re-arranged the stories to have the same format as the agenda component.

Added the following stories:
* Attribute stories for files, item-path, page size and file extensions.
* Templating
* Events
